feat: add solution modules tack
